backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
gunicorn==22.0.0
python-dotenv==1.0.1
supabase==2.5.0
web3==6.20.0
python-jose[cryptography]==3.3.0
httpx==0.31.0  # Updated version